Changed the frequency to 420.08mhz

+++ blaze-lite/core/lib/radio/Radio.cpp
@@ -18,7 +18,7 @@ void Radio::setCallSign(String sign) {
 /**
  * Initializes radio. Handles reset pin, frequency, and power config.
  */
-bool Radio::init(uint32_t freq) {
+bool Radio::init(float freq) {
   pinMode(rst_pin, OUTPUT);
   digitalWrite(rst_pin, LOW);
